Injecting OkHttpClient and Retrofit rather than building them inside class. Injecting fake retrofit instance during Espresso tests.
