+++ SuperHero/app/src/androidTest/kotlin/lt/setkus/superhero/app/espresso/MockApiActivityTestRule.kt
@@ -2,8 +2,11 @@ package lt.setkus.superhero.app.espresso
 
 import android.app.Activity
 import androidx.test.rule.ActivityTestRule
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.experimental.CoroutineCallAdapterFactory
 import org.koin.dsl.module.module
 import org.koin.standalone.StandAloneContext.loadKoinModules
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
 
 class MockApiActivityTestRule<T : Activity> : ActivityTestRule<T> {
     constructor(activityClass: Class<T>?) : super(activityClass, false, false)
@@ -11,8 +14,17 @@ class MockApiActivityTestRule<T : Activity> : ActivityTestRule<T> {
     override fun beforeActivityLaunched() {
         super.beforeActivityLaunched()
 
-        loadKoinModules(mockHttpModule)
+        loadKoinModules(
+            module {
+                single(override = true) { provideMockRetrofit() }
+            }
+        )
     }
-}
 
-val mockHttpModule = module {  }
+    private fun provideMockRetrofit() =
+        Retrofit.Builder()
+            .addConverterFactory(GsonConverterFactory.create())
+            .addCallAdapterFactory(CoroutineCallAdapterFactory())
+            .baseUrl("http://localhost:8080")
+            .build()
+}
