+++ SuperHero/app/src/main/kotlin/lt/setkus/superhero/app/di/module/HttpModule.kt
@@ -1,11 +1,15 @@
 package lt.setkus.superhero.app.di.module
 
 import lt.setkus.superhero.data.http.MarvelService
+import lt.setkus.superhero.data.http.provideOkHttpClient
+import lt.setkus.superhero.data.http.provideRetrofit
 import lt.setkus.superhero.utils.AppExecutors
 import org.koin.dsl.module.module
 
 val httpModule = module {
-    factory { MarvelService.getCharacterService() }
+    single { provideOkHttpClient() }
+    single { provideRetrofit(get()) }
+    single { MarvelService(get()) }
     single(name = "uiContext") { AppExecutors().uiContext }
     single(name = "networkContext") { AppExecutors().networkContext }
 }
