+++ SuperHero/app/src/main/kotlin/lt/setkus/superhero/data/heroes/SuperHeroesDataRepository.kt
@@ -1,17 +1,17 @@
 package lt.setkus.superhero.data.heroes
 
-import lt.setkus.superhero.data.http.CharacterService
+import lt.setkus.superhero.data.http.MarvelService
 import lt.setkus.superhero.domain.Result
 import lt.setkus.superhero.domain.heroes.SuperHero
 import lt.setkus.superhero.domain.heroes.SuperHeroesRepository
 import lt.setkus.superhero.utils.awaitResult
 
 class SuperHeroesDataRepository(
-    val service: CharacterService
+    val service: MarvelService
 ) : SuperHeroesRepository {
 
     override suspend fun loadSuperHeroes(): Result<List<SuperHero>> {
-        val result = service.getCharacters().awaitResult()
+        val result = service.getCharacterService().getCharacters().awaitResult()
         return when (result) {
             is Result.Error -> result
             is Result.Success -> flatMap(result.data)
