+++ SuperHero/app/src/main/kotlin/lt/setkus/superhero/data/http/RetrofitProvider.kt
@@ -1,0 +1,14 @@
+package lt.setkus.superhero.data.http
+
+import com.jakewharton.retrofit2.adapter.kotlin.coroutines.experimental.CoroutineCallAdapterFactory
+import okhttp3.OkHttpClient
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+fun provideRetrofit(client: OkHttpClient) =
+    Retrofit.Builder()
+        .client(client)
+        .addConverterFactory(GsonConverterFactory.create())
+        .addCallAdapterFactory(CoroutineCallAdapterFactory())
+        .baseUrl("https://gateway.marvel.com")
+        .build()
